% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessment_inventory.R
\name{assessment_inventory}
\alias{assessment_inventory}
\title{Obtain Study Information From Universal FQA Inventory as a Data Frame}
\usage{
assessment_inventory(data_set)
}
\arguments{
\item{data_set}{a data frame downloaded from Universal FQA using download_assessment() or directly from universalfqa.org}
}
\value{
A data frame with 9 columns:
\itemize{
   \item Scientific Name (character)
   \item Family (character)
   \item Acronym (character)
   \item Native? (character)
   \item C (numeric)
   \item W (numeric)
   \item Physiognomy (character)
   \item Duration (character)
   \item Common Name (character)
}
}
\description{
Obtain Study Information From Universal FQA Inventory as a Data Frame
}
\examples{
\dontrun{
## assessment_inventory can be used with a .csv file downloaded from the universal FQA website:

assessment_inventory(open_dunes)

## or with a download function:

assessment_inventory(download_assessment(25002))

## assessment_inventory can also be used with saved data from a download function:

df <- download_fqa(25002)
assessment_inventory(df)
}
}
