% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_datasets.R
\docType{data}
\name{chicago}
\alias{chicago}
\title{Overview Data of over 700 FQA Inventories (Chicago 2017 Database)}
\format{
A data frame with 725 rows and 52 variables
}
\usage{
data(chicago)
}
\value{
A data frame with 52 columns:
\itemize{
   \item Title (character)
   \item Date (POSIXct)
   \item Site Name (character)
   \item City (character)
   \item County (character)
   \item State (character)
   \item Country (character)
   \item FQA DB Region (character)
   \item FQA DB Publication Year (character)
   \item FQA DB Description (character)
   \item Practitioner (character)
   \item Latitude (character)
   \item Longitude (character)
   \item Weather Notes (character)
   \item Duration Notes (character)
   \item Community Type Notes (character)
   \item Other Notes (character)
   \item Private/Public (character)
   \item Total Mean C (numeric)
   \item Native Mean C (numeric)
   \item Total FQI (numeric)
   \item Native FQI (numeric)
   \item Adjusted FQI (numeric)
   \item % C value 0 (numeric)
   \item % C value 1-3 (numeric)
   \item % C value 4-6 (numeric)
   \item % C value 7-10 (numeric)
   \item Native Tree Mean C (numeric)
   \item Native Shrub Mean C (numeric)
   \item Native Herbaceous Mean C (numeric)
   \item Species Richness (numeric)
   \item Total Species (numeric)
   \item Native Species (numeric)
   \item Non-native Species (numeric)
   \item Species Wetness (numeric)
   \item Mean Wetness (numeric)
   \item Native Mean Wetness (numeric)
   \item Tree (numeric)
   \item Shrub (numeric)
   \item Vine (numeric)
   \item Forb (numeric)
   \item Grass (numeric)
   \item Sedge (numeric)
   \item Rush (numeric)
   \item Fern (numeric)
   \item Bryophyte (numeric)
   \item Annual (numeric)
   \item Perennial (numeric)
   \item Biennial (numeric)
   \item Native Annual (numeric)
   \item Native Perennial: (numeric)
   \item Native Biennial: (numeric)
}
}
\description{
A dataset containing the glance_fqa() data of 725 FQA inventories.
The variables are as follows:
}
\keyword{datasets}
