% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloading_functions.R
\name{download_assessment}
\alias{download_assessment}
\title{Download FQA by id number}
\usage{
download_assessment(assessment_id)
}
\arguments{
\item{assessment_id}{A numeric identifier of the desired floristic quality assessment, as specified by \href{https://universalfqa.org/}{universalfqa.org}. ID numbers for specified databases can be viewed with the \code{\link{download_fqa_assessments}} function.}
}
\value{
An untidy data frame in the original format of the Universal FQA website. Use \code{\link{assessment_glance}} for a tidy summary and \code{\link{assessment_inventory}} for raw species-level data.
}
\description{
Downloads a specified floristic quality assessment from \href{https://universalfqa.org/}{universalfqa.org}. ID numbers for assessments in various databases can be found using the \code{\link{download_fqa_assessments}} function.
}
\examples{
\dontrun{
databases <- download_fqa_databases()
# Note database 1 is the original 1994 Chicago edition.
chicago_assessments <- download_fqa_assessments(1) # Edison dune and swale has id number 25002.
edison <- download_assessment(25002)
edison_tidy <- glance_fqa(edison)
edison_species <- assess_fqa(edison)
}

}
