% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexing_functions.R
\name{download_fqa_databases}
\alias{download_fqa_databases}
\title{List all available floristic quality assessment databases}
\usage{
download_fqa_databases()
}
\value{
A data frame with 4 columns:
\itemize{
  \item database_id (numeric)
  \item region (character)
  \item year (numeric)
  \item description (character)
}
}
\description{
\code{download_fqa_databases} creates a data frame showing all public
floristic quality assessment databases available at
\href{https://universalfqa.org/}{universalfqa.org}.
}
\examples{
\dontrun{
databases <- download_fqa_databases()
}

}
