% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloading_functions.R
\name{download_transect}
\alias{download_transect}
\title{Download transect-level FQA data by id number}
\usage{
download_transect(transect_id)
}
\arguments{
\item{transect_id}{A numeric identifier of the desired transect-level floristic quality assessment, as specified by \href{https://universalfqa.org/}{universalfqa.org}. ID numbers for specified databases can be viewed with the \code{\link{download_fqa_transects}} function.}
}
\value{
An untidy data frame in the original format of the Universal FQA website. Use \code{\link{transect_glance}} for a tidy summary, \code{\link{transect_phys}} for a physiognometric overview, and \code{\link{transect_inventory}} for raw species-level data.
}
\description{
Downloads a specified transect-level floristic quality assessment from \href{https://universalfqa.org/}{universalfqa.org}. ID numbers for assessments in different databases can be found using \code{\link{download_fqa_assessments}}.
}
\examples{
\dontrun{
databases <- download_fqa_databases # Note database 1 is the original 1994 Chicago edition.
chicago_transects <- download_fqa_transects(1) # CBG Sand prairie swale fen A has id number 5932.
cbg <- download_transect(5932)
cbg_tidy <- transect_glance(edison)
cbg_species <- transect_inventory(edison)
cbg_phys <- transect_phys(edison)
}

}
