% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transect_glance.R
\name{transect_glance}
\alias{transect_glance}
\title{Obtain tidy summary information for a floristic quality transect assessment}
\usage{
transect_glance(data_set)
}
\arguments{
\item{data_set}{A data set downloaded from
\href{https://universalfqa.org/}{universalfqa.org} either manually or using
\code{\link[=download_transect]{download_transect()}}.}
}
\value{
A data frame with 1 row and 54 columns:
\itemize{
   \item title (character)
   \item date (date)
   \item site_name (character)
   \item city (character)
   \item county (character)
   \item state (character)
   \item country (character)
   \item omernik_level_three_ecoregion (character)
   \item fqa_db_region (character)
   \item fqa_db_publication_year (character)
   \item fqa_db_description (character)
   \item fqa_db_selection_name (character)
   \item custom_fqa_db_name (character)
   \item custom_fqa_db_description (character)
   \item practitioner (character)
   \item latitude (character)
   \item longitude (character)
   \item community_code (character)
   \item community_name (character)
   \item community_type_notes (character)
   \item weather_notes (character)
   \item duration_notes (character)
   \item environment_description (character)
   \item other_notes (character)
   \item transect_plot_type (character)
   \item plot_size (numeric) Plot size in square meters
   \item quadrat_subplot_size (numeric) Quadrat or subplot size in square meters
   \item transect_length (numeric) Transect length in meters
   \item sampling_design_description (character)
   \item cover_method (character)
   \item private_public (character)
   \item total_mean_c (numeric)
   \item cover_weighted_mean_c (numeric)
   \item native_mean_c (numeric)
   \item total_fqi (numeric)
   \item native_fqi (numeric)
   \item cover_weighted_fqi (numeric)
   \item cover_weighted_native_fqi (numeric)
   \item adjusted_fqi (numeric)
   \item c_value_zero (numeric) Percent of c-values 0
   \item c_value_low (numeric) Percent of c-values 1-3
   \item c_value_mid (numeric) Percent of c-values 4-6
   \item c_value_high (numeric) Percent of c-values 7-10
   \item total_species (numeric)
   \item native_species (numeric)
   \item non_native_species (numeric)
   \item mean_wetness (numeric)
   \item native_mean_wetness (numeric)
   \item annual (numeric)
   \item perennial (numeric)
   \item biennial (numeric)
   \item native_annual (numeric)
   \item native_perennial (numeric)
   \item native_biennial (numeric)
}
}
\description{
\code{transect_glance()} tidies a floristic quality transect assessment data
set obtained from \href{https://universalfqa.org/}{universalfqa.org}.
}
\examples{
# While transect_glance can be used with a .csv file downloaded manually
# from the universal FQA website, it is most typically used in combination
# with download_transect().

tyler <- download_transect(6352)
transect_glance(tyler)

}
