% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transect_inventory.R
\name{transect_inventory}
\alias{transect_inventory}
\title{Obtain Species Information From a Universal FQA Transect as a Data Frame}
\usage{
transect_inventory(data_set)
}
\arguments{
\item{data_set}{a data frame downloaded from Universal FQA using download_transect() or other similar function}
}
\value{
A data frame with 6 columns:
\itemize{
  \item Species (character)
  \item Family (character)
  \item Acronym (character)
  \item Nativity (character)
  \item C (numeric)
  \item W (numeric)
  \item Physiognomy (character)
  \item Duration (character)
  \item Frequency (numeric)
  \item Coverage (numeric)
  \item Relative Frequency % (numeric)
  \item Relative Coverage % (numeric)
  \item Relative Importance Value (numeric)
}
}
\description{
Obtain Species Information From a Universal FQA Transect as a Data Frame
}
\examples{
## transect_inventory can be used with a download function:

transect_inventory(download_transect(6325))

## transect_inventory can also be used with saved data from a download function:

df <- download_transect(6325)
transect_inventory(df)
}
