% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessment_glance.R
\name{assessment_glance}
\alias{assessment_glance}
\title{Obtain tidy summary information for a floristic quality assessment}
\usage{
assessment_glance(data_set)
}
\arguments{
\item{data_set}{A data set downloaded from
\href{https://universalfqa.org/}{universalfqa.org} either manually or using
\code{\link[=download_assessment]{download_assessment()}}}
}
\value{
A data frame with 53 columns:
\itemize{
   \item assessment_id (numeric)
   \item title (character)
   \item date (date)
   \item site_name (character)
   \item city (character)
   \item county (character)
   \item state (character)
   \item country (character)
   \item fqa_db_region (character)
   \item fqa_db_publication_year (character)
   \item fqa_db_description (character)
   \item custom_fqa_db_name (character)
   \item custom_fqa_db_description (character)
   \item practitioner (character)
   \item latitude (character)
   \item longitude (character)
   \item weather_notes (character)
   \item duration_notes (character)
   \item community_type_notes (character)
   \item other_notes (character)
   \item private_public (character)
   \item total_mean_c (numeric)
   \item native_mean_c (numeric)
   \item total_fqi (numeric)
   \item native_fqi (numeric)
   \item adjusted_fqi (numeric)
   \item c_value_zero (numeric) Percent of c-values 0
   \item c_value_low (numeric) Percent of c-values 1-3
   \item c_value_mid (numeric) Percent of c-values 4-6
   \item c_value_high (numeric) Percent of c-values 7-10
   \item native_tree_mean_c (numeric)
   \item native_shrub_mean_c (numeric)
   \item native_herbaceous_mean_c (numeric)
   \item total_species (numeric)
   \item native_species (numeric)
   \item non_native_species (numeric)
   \item mean_wetness (numeric)
   \item native_mean_wetness (numeric)
   \item tree (numeric)
   \item shrub (numeric)
   \item vine (numeric)
   \item forb (numeric)
   \item grass (numeric)
   \item sedge (numeric)
   \item rush (numeric)
   \item fern (numeric)
   \item bryophyte (numeric)
   \item annual (numeric)
   \item perennial (numeric)
   \item biennial (numeric)
   \item native_annual (numeric)
   \item native_perennial (numeric)
   \item native_biennial (numeric)
}
}
\description{
\code{assessment_glance()} tidies a floristic quality assessment data set
obtained from \href{https://universalfqa.org/}{universalfqa.org}.
}
\examples{
# While assessment_glance can be used with a .csv file downloaded manually
# from the universal FQA website, it is most typically used in combination
# with download_assessment().

edison <- download_assessment(25002)
assessment_glance(edison)

}
