% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessment_inventory.R
\name{assessment_inventory}
\alias{assessment_inventory}
\title{Obtain species details for a given floristic quality assessment}
\usage{
assessment_inventory(data_set)
}
\arguments{
\item{data_set}{a data frame downloaded from Universal FQA using download_assessment() or directly from universalfqa.org}
}
\value{
A data frame with 9 columns:
\itemize{
   \item Scientific Name (character)
   \item Family (character)
   \item Acronym (character)
   \item Native? (character)
   \item C (numeric)
   \item W (numeric)
   \item Physiognomy (character)
   \item Duration (character)
   \item Common Name (character)
}
}
\description{
Obtain species details for a given floristic quality assessment
}
\examples{
\dontrun{
# While assessment_glance can be used with a .csv file downloaded
# manually from the universal FQA website, it is most typically used
# in combination with \code{\link{download_assessment}}:

edison <- download_assessment(25002)
assessment_inventory(edison)
}

}
