% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexing_functions.R
\name{download_fqa_databases}
\alias{download_fqa_databases}
\title{Obtain a data frame of all existing FQA databases}
\usage{
download_fqa_databases()
}
\value{
A data frame with four columns:
\itemize{
  \item database_id (numeric)
  \item region (character)
  \item year (numeric)
  \item description (character)
}
}
\description{
Downloads a data frame of public floristic quality assessment databases from \href{https://universalfqa.org/}{universalfqa.org}.
}
\examples{
\dontrun{
databases <- download_fqa_databases()
}

}
