% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexing_functions.R
\name{download_fqa_transects}
\alias{download_fqa_transects}
\title{Obtain a data frame of all available public FQA transect assessments}
\usage{
download_fqa_transects(database_id)
}
\arguments{
\item{database_id}{A numeric identifier of the desired database, as specified by \href{https://universalfqa.org/}{universalfqa.org}. The id numbers can be viewed with the \code{\link{download_fqa_databases}} function.}
}
\value{
A data frame with 5 columns:
\itemize{
  \item id (numeric)
  \item assessment (character)
  \item date (date)
  \item site (character)
  \item practitioner (character)
  }
}
\description{
Downloads a data frame of publicly-available transect assessments for a given FQA database from \href{https://universalfqa.org/}{universalfqa.org}.
Databases should be specified by their \code{database_id} number, which can be obtained with the \code{\link{download_fqa_databases}} function.
}
\examples{
\dontrun{
databases <- download_fqa_databases()
# Note that the most current Chicago database has id_number 149.
chicago_2017_transects <- download_fqa_assessments(149)
}

}
