% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloading_functions.R
\name{download_transect_list}
\alias{download_transect_list}
\title{Download multiple floristic quality transect assessments}
\usage{
download_transect_list(database, ...)
}
\arguments{
\item{database}{A numeric identifier of the desired floristic quality assessment database, as specified by \href{https://universalfqa.org/}{universalfqa.org}. Database id numbers can be viewed with the \code{\link{download_fqa_databases}} function.}

\item{...}{`dplyr`-style filtering criteria for the desired transect-level assessments. The following variables may be used:
\itemize{
  \item id (numeric)
  \item assessment (character)
  \item date (POSIXct)
  \item site (character)
  \item practitioner (character)
}}
}
\value{
A list of data frames matching the given search criteria. Each is an untidy data frame in the original format of the Universal FQA website. Use \code{\link{transect_list_glance}} for a tidy summary.
}
\description{
Searches a specified FQA database for transect assessments matching given criteria, then downloads all the matches from \href{https://universalfqa.org/}{universalfqa.org}. Download speeds from that website may be slow, causing delays in the evaluation of this function.
}
\examples{
\dontrun{
databases <- download_fqa_databases()
# Note database 1 is the original 1994 Chicago edition.
chicago_transects <- download_fqa_transects(1)
fcna <- download_transect_list(site == "FCNA")
}

}
