% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexing_functions.R
\name{index_fqa_databases}
\alias{index_fqa_databases}
\title{List all available floristic quality assessment databases}
\usage{
index_fqa_databases()
}
\value{
A data frame with 4 columns:
\itemize{
  \item database_id (numeric)
  \item region (character)
  \item year (numeric)
  \item description (character)
}
}
\description{
\code{index_fqa_databases} produces a data frame showing all floristic
quality assessment databases publicly available at
\href{https://universalfqa.org/}{universalfqa.org}.
}
\examples{
\dontrun{
databases <- index_fqa_databases()
}

}
