% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transect_inventory.R
\name{transect_inventory}
\alias{transect_inventory}
\title{Obtain species details for a given floristic quality transect assessment}
\usage{
transect_inventory(data_set)
}
\arguments{
\item{data_set}{a data frame downloaded from Universal FQA using download_transect() or other similar function}
}
\value{
A data frame with 6 columns:
\itemize{
  \item Species (character)
  \item Family (character)
  \item Acronym (character)
  \item Nativity (character)
  \item C (numeric)
  \item W (numeric)
  \item Physiognomy (character)
  \item Duration (character)
  \item Frequency (numeric)
  \item Coverage (numeric)
  \item Relative Frequency % (numeric)
  \item Relative Coverage % (numeric)
  \item Relative Importance Value (numeric)
}
}
\description{
Obtain species details for a given floristic quality transect assessment
}
\examples{
\dontrun{
## while transect_glance can be used with a .csv file downloaded manually from the universal FQA website, it is most typically used in combination with \code{\link{download_transect}}:

tyler <- download_transect(6352)
transect_glance(tyler)
}

}
