% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transect_phys.R
\name{transect_phys}
\alias{transect_phys}
\title{Obtain Physiognomy Information From a Universal FQA Transect as a Data Frame}
\usage{
transect_phys(data_set)
}
\arguments{
\item{data_set}{a data frame downloaded from Universal FQA using download_transect() or other similar function}
}
\value{
A data frame with 6 columns:
\itemize{
   \item Physiognomy (character)
   \item Frequency (numeric)
   \item Coverage (numeric)
   \item Relative Frequency % (numeric)
   \item Relative Coverage % (numeric)
   \item Relative Importance Value (numeric)
}
}
\description{
Obtain Physiognomy Information From a Universal FQA Transect as a Data Frame
}
\examples{
\dontrun{
## transect_phys can be used with a download function:

transect_phys(download_transect(6325))

## transect_phys can also be used with saved data from a download function:

df <- download_transect(6325)
transect_phys(df)
}
}
